+++ frontend/lib/widgets/ui/doctor/doctor_exam_question_rule_edit.dart
@@ -1,23 +1,19 @@
-import 'dart:async';
 import 'dart:math';
-import 'dart:ui' as ui;
 
-import 'package:aphasia_recovery/enum/command_actions.dart';
 import 'package:aphasia_recovery/mixin/widgets_mixin.dart';
-import 'package:aphasia_recovery/utils/algorithm.dart';
 import 'package:aphasia_recovery/utils/io/assets.dart';
 import 'package:aphasia_recovery/utils/http/http_common.dart';
 import 'package:aphasia_recovery/widgets/ui/common/common.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:aphasia_recovery/mixin/eval_rule_mixin.dart';
 import 'doctor_choice_setting_dialog.dart';
+import 'doctor_command_question_action_record_page.dart';
+import 'doctor_item_finding_question_area_setting_page.dart';
 import 'doctor_item_slot_edit_dialog.dart';
 import 'doctor_question_score_condition_edit_dialog.dart';
 import '../../../models/question/question.dart';
 import '../../../models/rules.dart';
 import '../../../utils/common_widget_function.dart';
-import '../do_exam/command_question.dart';
 
 
 class DoctorExamQuestionRuleEditSubPage extends StatefulWidget {
@@ -785,460 +781,5 @@ class _DoctorExamQuestionRuleEditSubPageState extends State<DoctorExamQuestionRu
         ],
       ),
     ));
-  }
-}
-
-class CommandQuestionActionRecordPage extends StatefulWidget {
-  final EvalCommandQuestionByCorrectActionCount currRule;
-  const CommandQuestionActionRecordPage({super.key, required this.currRule,});
-
-  @override
-  State<CommandQuestionActionRecordPage> createState() => _CommandQuestionActionRecordPageState();
-}
-
-class _CommandQuestionActionRecordPageState extends State<CommandQuestionActionRecordPage> with UseCommonStyles {
-  late EvalCommandQuestionByCorrectActionCount currRule;
-  late List<ItemSlot> currSlots;
-  late List<StackableItemSlot> slots;
-  late List<CommandActions> actions;
-  late String commandText;
-  CommandActions? currAction;
-
-  void resetStates() {
-    currRule = widget.currRule;
-    currSlots = currRule.slots.map((e) => e.copy()).toList();
-    slots = currSlots.map((e) => e.itemName == null? StackableItemSlot() : StackableItemSlot(e)).toList();
-    actions = [];
-    commandText = "";
-  }
-
-  @override
-  void initState() {
-    resetStates();
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (currRule != widget.currRule) {
-      resetStates();
-    }
-
-    initStyles(context);
-
-    return Scaffold(
-      appBar: AppBar(leading: const BackButton(), title: Text("拖动物体进行操作录制", style: commonStyles?.titleStyle,)),
-      body: SafeArea(
-        child: Center(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 8.0),
-            child: SingleChildScrollView(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  _buildItemSlots(context, currRule),
-                  const SizedBox(height: 16,),
-                  Text("操作对应的指令文本：$commandText", style: commonStyles?.bodyStyle,),
-                  const SizedBox(height: 16,),
-                  Row(
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      ElevatedButton(
-                        onPressed: () {
-                          confirm(context, title: "确认", body: "确定要重新录制吗？", commonStyles: commonStyles,
-                            onConfirm: (context) {
-                              Navigator.pop(context);
-                              setState(() {
-                                resetStates();
-                              });
-                            }
-                          );
-                        },
-                        child: Text("重新录制", style: commonStyles?.bodyStyle,),
-                      ),
-                      const SizedBox(width: 16,),
-                      ElevatedButton(
-                        onPressed: () {
-                          if (actions.isEmpty) {
-                            toast(context, msg: "请至少记录一个动作再完成录制", btnText: "确认");
-                            return;
-                          }
-
-                          if (currRule.actions.isNotEmpty) {
-                            confirm(context, title: "确认", body: "当前规则已录制过正确操作，继续将覆盖已录制的操作，确认继续吗？", commonStyles: commonStyles,
-                              onConfirm: (context) {
-                                // 关闭确认弹窗
-                                Navigator.pop(context);
-
-                                // 从录制页返回到上一页
-                                Navigator.pop(context, actions);
-                              }
-                            );
-                          } else {
-                            Navigator.pop(context, actions);
-                          }
-                        },
-                        style: ElevatedButton.styleFrom(backgroundColor: commonStyles?.primaryColor),
-                        child: Text("完成录制", style: commonStyles?.bodyStyle?.copyWith(color: commonStyles?.onPrimaryColor),),
-                      ),
-                    ],
-                  ),
-
-                ],
-              ),
-            ),
-          ),
-        ),
-      )
-    );
-  }
-
-  Widget _buildItemSlots(BuildContext context, EvalCommandQuestionByCorrectActionCount rule) {
-    final media = MediaQuery.of(context);
-    final screenAspectRatio = media.size.aspectRatio;
-    // debugPrint(screenAspectRatio.toString());
-    const spacing = 2.0;
-
-    return GridView.count(
-        crossAxisCount: 5,
-        mainAxisSpacing: spacing,
-        crossAxisSpacing: spacing,
-        childAspectRatio: screenAspectRatio,
-        shrinkWrap: true,
-        children: slots.asMap().entries.map((e) => Container(
-          decoration: BoxDecoration(border: Border.all(color: Colors.grey)),
-          child: Builder(
-              builder: (context) {
-                final index = e.key;
-                final slot = e.value;
-
-                Widget content;
-                if (slot.items.isNotEmpty) {
-                  content = LayoutBuilder(
-                    builder: (context, constraints) {
-                      return Draggable<StackableItemSlot>(
-                        data: slot,
-                        onDragStarted: () {
-                          currAction = CommandActions(sourceSlotIndex: index, firstAction: ClickAction.take);
-                        },
-                        onDraggableCanceled: (v, offset) {
-                          currAction = null;
-                        },
-                        feedback: SizedBox(
-                          width: constraints.maxWidth,
-                          height: constraints.maxHeight,
-                          child: _buildDragFeedback(slot),
-                        ),
-                        childWhenDragging: _buildSlotImagesWhenDragging(slot),
-                        child: InkWell(
-                          onTap: () {
-                            setState(() {
-                              actions.add(CommandActions(sourceSlotIndex: index, firstAction: ClickAction.touch));
-                              commandText = rule.generateCommandTextFromActions(actions);
-                            });
-                          },
-                          child: _buildSlotImages(slot)
-                        ),
-                      );
-                    }
-                  );
-                } else {
-                  content = const SizedBox.shrink();
-                }
-                return DragTarget(
-                  builder: (BuildContext context, List<Object?> candidateData, List<dynamic> rejectedData) {
-                    return content;
-                  },
-                  onAccept: (StackableItemSlot incomingSlot) {
-                    assert(incomingSlot.items.isNotEmpty && currAction?.sourceSlotIndex != null);
-                    final actionType = slot.items.isNotEmpty ? PutDownAction.putDown : PutDownAction.cover;
-
-                    setState(() {
-                      actions.add(currAction!..setSecondAction(index, actionType));
-                      currAction = null;
-
-                      commandText = rule.generateCommandTextFromActions(actions);
-                      slot.pushItem(incomingSlot.popItem()!);
-                    });
-                  },
-                );
-              }
-          ),
-        )).toList()
-    );
-  }
-
-  Widget _buildSlotImages(StackableItemSlot slot) {
-    return Center(
-      child: Stack(
-        alignment: Alignment.center,
-        children: slot.items.map((item) {
-          return item.itemImageUrl != null ? Image.network(item.itemImageUrl!,
-            fit: BoxFit.contain,
-          ) : Image.asset(item.itemImageAssetPath!,
-            fit: BoxFit.contain,
-          );
-        }).toList(),
-      ),
-    );
-  }
-
-  Widget _buildSlotImagesWhenDragging(StackableItemSlot slot) {
-    return Center(
-      child: Stack(
-        children: slot.items.asMap().entries.map((e) {
-          var index = e.key;
-          var item = e.value;
-
-          if (index == slot.items.length - 1) {
-            return const SizedBox.shrink();
-          }
-
-          return item.itemImageUrl != null ? Image.network(item.itemImageUrl!,
-            fit: BoxFit.contain,
-          ) : Image.asset(item.itemImageAssetPath!,
-            fit: BoxFit.contain,
-          );
-        }).toList(),
-      ),
-    );
-  }
-
-  Widget _buildDragFeedback(StackableItemSlot slot) {
-    assert(slot.items.isNotEmpty);
-    final item = slot.items.last;
-
-    return Center(
-      child: item.itemImageUrl != null ? Image.network(item.itemImageUrl!,
-        fit: BoxFit.contain,
-      ) : Image.asset(item.itemImageAssetPath!,
-        fit: BoxFit.contain,
-      )
-    );
-  }
-}
-
-class ItemFindingQuestionAreaSettingPage extends StatefulWidget {
-  final ItemFindingQuestion question;
-  const ItemFindingQuestionAreaSettingPage({super.key, required this.question});
-
-  @override
-  State<ItemFindingQuestionAreaSettingPage> createState() => _ItemFindingQuestionAreaSettingPageState();
-}
-
-class _ItemFindingQuestionAreaSettingPageState extends State<ItemFindingQuestionAreaSettingPage> with UseCommonStyles {
-  static const int maxPointCount = 20;
-  late ItemFindingQuestion currQuestion;
-
-  int pointCount = 0;
-
-  late List<List<double>> coordinates;
-  bool convexHullFinished = false;
-
-  void resetState() {
-    clearState();
-    final rule = currQuestion.evalRule as EvalItemFoundQuestion;
-    pointCount = rule.coordinates.isEmpty ? 0 : maxPointCount - rule.coordinates.length;
-
-    coordinates = rule.coordinates.isEmpty ? [] : []..addAll(rule.coordinates);
-  }
-
-  void clearState() {
-    currQuestion = widget.question;
-
-    pointCount = 0;
-
-    coordinates = [];
-    convexHullFinished = false;
-  }
-
-  void addCoordinate(double px, double py, double maxX, double maxY) {
-    if (pointCount == maxPointCount) {
-      toast(context, msg: "至多指定20个顶点，当前顶点数量已达上限。", btnText: "确认");
-      return;
-    }
-
-    coordinates.add(normalizePosition(px, py, maxX, maxY));
-    pointCount++;
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    resetState();
-  }
-
-  Future<ui.Image> getImageFromProvider(ImageProvider imageProvider) async {
-    Completer<ui.Image> completer = Completer<ui.Image>();
-    imageProvider.resolve(const ImageConfiguration()).addListener(
-      ImageStreamListener((ImageInfo info, bool _) async {
-        ByteData? byteData = await info.image.toByteData();
-        Uint8List? uint8List = byteData?.buffer.asUint8List();
-        ui.Codec codec = await ui.instantiateImageCodec(uint8List!);
-        ui.FrameInfo frameInfo = await codec.getNextFrame();
-        completer.complete(frameInfo.image);
-      }),
-    );
-    return completer.future;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    initStyles(context);
-
-    final imageCompleter = Completer<ui.Image>();
-    Image questionImage;
-    if (isImageUrlAssets(currQuestion.imageUrl)) {
-      questionImage = Image(image: AssetImage(currQuestion.imageUrl!), fit: BoxFit.contain,);
-    } else {
-      questionImage = Image(
-        image: NetworkImage(currQuestion.imageUrl!),
-        fit: BoxFit.contain,
-      );
-    }
-
-    questionImage.image.resolve(const ImageConfiguration())
-        .addListener(ImageStreamListener((imageInfo, _) {
-      imageCompleter.complete(imageInfo.image);
-    }));
-
-    if (currQuestion != widget.question) {
-      resetState();
-    }
-
-
-    return Scaffold(
-      appBar: AppBar(leading: const BackButton(), title: Text("场景寻物题点击区域设置", style: commonStyles?.titleStyle,)),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 32),
-          child: SingleChildScrollView(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text("仅支持设置凸多边形区域，请在图片上点击设置多边形区域的顶点（至多20个顶点），设置完毕后点击“完成”按钮以生成区域，系统会自动根据设置的顶点求出一个最大的凸多边形区域，区域生成后点击“确认设置”按钮保存并返回到规则设置页面", style: commonStyles?.bodyStyle,),
-                Text("顶点数：$pointCount/$maxPointCount", style: commonStyles?.bodyStyle,),
-                const Divider(height: 32,),
-                FutureBuilder<ui.Image>(
-                  future: imageCompleter.future,
-                  builder: (context, snapshot) {
-                    if (snapshot.hasError) {
-                      toast(context, msg: "图片加载失败，请重试。", btnText: "确认");
-                      return Center(
-                        child: Text("加载中，请稍候", style: commonStyles?.hintTextStyle,),
-                      );
-                    } else if (!snapshot.hasData) {
-                      return Center(
-                        child: Text("加载中，请稍候", style: commonStyles?.hintTextStyle,),
-                      );
-                    }
-
-                    final image = snapshot.data!;
-                    final mediaSize = MediaQuery.of(context).size;
-                    double maxWidth = mediaSize.width * 0.7;
-                    double maxHeight = mediaSize.height * 0.7;
-                    double boxWidth;
-                    double boxHeight;
-                    if (image.height * (maxWidth / image.width) <= maxHeight) {
-                      boxWidth = maxWidth;
-                      boxHeight = image.height * (maxWidth / image.width);
-                    } else {
-                      boxWidth = image.width * (maxHeight / image.height);
-                      boxHeight = maxHeight;
-                    }
-
-                    return Container(
-                      width: boxWidth,
-                      height: boxHeight,
-                      decoration: BoxDecoration(
-                        border: Border.all(width: 1.0)
-                      ),
-                      child: Stack(
-                        fit: StackFit.expand,
-                        children: [
-                          GestureDetector(
-                            onTapDown: (details) {
-                              final RenderBox box = context.findRenderObject() as RenderBox;
-                              // find the coordinate
-                              final Offset localOffset = box.globalToLocal(details.globalPosition);
-                              final posX = localOffset.dx;
-                              final posY = localOffset.dy;
-                              setState(() {
-                                addCoordinate(posX, posY, boxWidth, boxHeight);
-                              });
-                            },
-                            child: questionImage
-                          ),
-                          ...coordinates.map((e) => Positioned(
-                              left: e.first * boxWidth - 9,
-                              top: e.last * boxHeight - 9,
-                              width: 18,
-                              height: 18,
-                              // child: Text("${(e.first * 1000).roundToDouble() / 1000};${(e.last * 1000).roundToDouble() / 1000}",)
-                              child: Container(
-                                decoration: BoxDecoration(
-                                  border: Border.all(color: Colors.black, width: 2),
-                                  shape: BoxShape.circle,
-                                ),
-                                child: const Center(
-                                  child: Icon(Icons.circle_rounded, color: Colors.green, size: 12.0,)
-                                ),
-                              ),
-                          )).toList(),
-                        ],
-                      ),
-                    );
-                  },
-                ),
-                const Divider(height: 32,),
-                Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      ElevatedButton(
-                        onPressed: () {
-                          setState(() {
-                            clearState();
-                          });
-                        },
-                        child: Text("重设", style: commonStyles?.bodyStyle,)
-                      ),
-                      const SizedBox(width: 16,),
-                      ElevatedButton(
-                        onPressed: () {
-                          setState(() {
-                            if (coordinates.length < 3) {
-                              toast(context, msg: "请至少设置3个顶点，至少3个顶点才能形成一个封闭区域", btnText: "确认");
-                              return;
-                            }
-                            coordinates = convexHull(coordinates);
-                            // debugPrint(coordinates.toString());
-                            convexHullFinished = true;
-                          });
-                        },
-                        child: Text("完成", style: commonStyles?.bodyStyle,)
-                      ),
-                      const SizedBox(width: 16,),
-                      ElevatedButton(
-                        onPressed: () {
-                          if (convexHullFinished) {
-                            Navigator.pop(context, coordinates);
-                          } else {
-                            toast(context, msg: "请先点击“完成”按钮生成区域再确认设置", btnText: "确认");
-                          }
-                        },
-                        child: Text("确认设置", style: commonStyles?.bodyStyle,)
-                      ),
-                    ],
-                  ),
-                )
-              ],
-            ),
-          ),
-        )
-      ),
-    );
   }
 }
